enable on-device builds
Partial compatibility for on-device builds.

There is no guarantee that it will be possible to build all available
packages and built packages will have same reliability that cross-compiled
but should solve "self-hosting" problems as much as possible.

+++ scripts/build/setup/termux_setup_cmake.sh
@@ -5,15 +5,33 @@ termux_setup_cmake() {
 	local TERMUX_CMAKE_TARNAME=cmake-${TERMUX_CMAKE_VERSION}-Linux-x86_64.tar.gz
 	local TERMUX_CMAKE_TARFILE=$TERMUX_PKG_TMPDIR/$TERMUX_CMAKE_TARNAME
 	local TERMUX_CMAKE_FOLDER=$TERMUX_COMMON_CACHEDIR/cmake-$TERMUX_CMAKE_VERSION
-	if [ ! -d "$TERMUX_CMAKE_FOLDER" ]; then
-		termux_download https://cmake.org/files/v$TERMUX_CMAKE_MAJORVESION/$TERMUX_CMAKE_TARNAME \
+
+	if [ -z "$TERMUX_ON_DEVICE_BUILD" ]; then
+		if [ ! -d "$TERMUX_CMAKE_FOLDER" ]; then
+			termux_download https://cmake.org/files/v$TERMUX_CMAKE_MAJORVESION/$TERMUX_CMAKE_TARNAME \
 				"$TERMUX_CMAKE_TARFILE" \
 				757e7ddd29d11422334a60d7cd9c19b7af20349a03cb15aa3c879f1b54193c95
-		rm -Rf "$TERMUX_PKG_TMPDIR/cmake-${TERMUX_CMAKE_VERSION}-Linux-x86_64"
-		tar xf "$TERMUX_CMAKE_TARFILE" -C "$TERMUX_PKG_TMPDIR"
-		mv "$TERMUX_PKG_TMPDIR/cmake-${TERMUX_CMAKE_VERSION}-Linux-x86_64" \
-		   "$TERMUX_CMAKE_FOLDER"
+			rm -Rf "$TERMUX_PKG_TMPDIR/cmake-${TERMUX_CMAKE_VERSION}-Linux-x86_64"
+			tar xf "$TERMUX_CMAKE_TARFILE" -C "$TERMUX_PKG_TMPDIR"
+			mv "$TERMUX_PKG_TMPDIR/cmake-${TERMUX_CMAKE_VERSION}-Linux-x86_64" \
+				"$TERMUX_CMAKE_FOLDER"
+		fi
+
+		export PATH=$TERMUX_CMAKE_FOLDER/bin:$PATH
+	else
+		if [ "$(dpkg-query -W -f '${db:Status-Status}\n' cmake 2>/dev/null)" != "installed" ]; then
+			echo "Package 'cmake' is not installed."
+			echo "You can install it with"
+			echo
+			echo "  pkg install cmake"
+			echo
+			echo "or build it from source with"
+			echo
+			echo "  ./build-package.sh cmake"
+			echo
+			exit 1
+		fi
 	fi
-	export PATH=$TERMUX_CMAKE_FOLDER/bin:$PATH
+
 	export CMAKE_INSTALL_ALWAYS=1
 }
